requirements.txt
# python requirements for OpenDataPedia
jinja2
loguru
markdown
pyyaml
typing
unidecode